token info data is now prefetched
