+++ .env
@@ -1,10 +1,10 @@
 
 # DFX CANISTER ENVIRONMENT VARIABLES
 DFX_VERSION='0.27.0'
-DFX_NETWORK='local'
-CANISTER_ID_TRABYTERHUB_FRONTEND='u6s2n-gx777-77774-qaaba-cai'
-CANISTER_ID_TRABYTERHUB_BACKEND='uxrrr-q7777-77774-qaaaq-cai'
-CANISTER_ID_INTERNET_IDENTITY='uzt4z-lp777-77774-qaabq-cai'
-CANISTER_ID='uzt4z-lp777-77774-qaabq-cai'
-CANISTER_CANDID_PATH='/home/ghostux/dev_github/5000Slices/webpage/.dfx/local/canisters/internet_identity/internet_identity.did'
+DFX_NETWORK='ic'
+CANISTER_ID_TRABYTERHUB_BACKEND='c33rl-3yaaa-aaaap-qp3bq-cai'
+CANISTER_ID_TRABYTERHUB_FRONTEND='c42x7-waaaa-aaaap-qp3ba-cai'
+CANISTER_ID_INTERNET_IDENTITY='rdmx6-jaaaa-aaaaa-aaadq-cai'
+CANISTER_ID='c33rl-3yaaa-aaaap-qp3bq-cai'
+CANISTER_CANDID_PATH='/home/ghostux/dev_github/5000Slices/webpage/.dfx/ic/canisters/TrabyterHub_backend/TrabyterHub_backend.did'
 # END DFX CANISTER ENVIRONMENT VARIABLES
